GEODE-5971 Refactor Lucene Index commands to support ResultModel and GfshCommand
