+++ geode-lucene/src/main/resources/META-INF/services/org.springframework.shell.core.CommandMarker
@@ -1,2 +1,6 @@
 # Lucene Extensions command
-org.apache.geode.cache.lucene.internal.cli.LuceneIndexCommands
+org.apache.geode.cache.lucene.internal.cli.LuceneCreateIndexCommand
+org.apache.geode.cache.lucene.internal.cli.LuceneDescribeIndexCommand
+org.apache.geode.cache.lucene.internal.cli.LuceneDestroyIndexCommand
+org.apache.geode.cache.lucene.internal.cli.LuceneListIndexCommand
+org.apache.geode.cache.lucene.internal.cli.LuceneSearchIndexCommand
